requirements.txt
torch
torchvision
gin-config